generate auth tokens for rsvp form submissions
The idea here is to store the document key for the rsvp datastore record,
then encrypt it as a json web token to be stored in the rsvp form as a
hidden input. We use the same private key to sign this token as is used
to authenticate against GCP datastore, which is probably fine for our use
here.