refactor main into multiple modules
Note there is still some refactoring that should happen here -- in
particular, a number of the RsvpService methods should live in the
datastore::client module, and the structs in main.rs should be split into
some sort of auth module.

We should also audit the way public fields and methods are being exposed,
there is likely a better way to structure some of these.